third_party/blink/web_tests/fast/xmlhttprequest/xmlhttprequest-responsetype-set-at-headers-received-expected.txt

This tests that the XMLHttpRequest responseType attribute is modifiable in the HEADERS_RECEIVED state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

xhr.responseType = 'document';
PASS xhr.responseType = 'text'; threw exception InvalidStateError: Failed to set the 'responseType' property on 'XMLHttpRequest': The response type cannot be set if the object's state is LOADING or DONE..
PASS xhr.responseType = 'text'; threw exception InvalidStateError: Failed to set the 'responseType' property on 'XMLHttpRequest': The response type cannot be set if the object's state is LOADING or DONE..
PASS xhr.response === xhr.responseXML is true
PASS xhr.responseType is "document"
PASS successfullyParsed is true

TEST COMPLETE
